test(skills): cover BOM handling and split the prose suite by concern
Adds the BOM cases, file and stdin, with the no-BOM file as the negative
control, mirroring the two the fence suite already carries.

That pushed the file past the 500-line ceiling, so it splits along the
seam it already had: test_prose_lint.py keeps the detectors, and
test_prose_lint_cli.py takes the exit-code contract, rule discovery, and
output formats. Splitting a test file costs nothing a vendored script
would, which is why the ceiling gets an extraction here and an inline
exemption there.
